+++ src/config.py
@@ -1,7 +1,7 @@
 import traceback, ConfigParser, ast, time, os, sys
 # configuration
 DEFAULT_CONFIG = """[General]
-command = java -jar minecraft_server.1.7.10.jar nogui
+command = java -jar minecraft_server.1.8.jar nogui
 auto-restart = True
 pre-1.7-mode = False
 timed-reboot = False
@@ -48,6 +48,7 @@ proxy-bind = 0.0.0.0
 server-port = 25564
 motd = Minecraft Server
 online-mode = True
+max-players = 1024
 """
 
 """[Web]
@@ -59,7 +60,7 @@ password = blahblah98
 public-stats = True"""
 
 class Config:
-	version = "0.7.1"
+	version = "0.7.2"
 	buildType = "dev" # dev, beta, or release
 	debug = False
 	def __init__(self, log):
@@ -78,7 +79,7 @@ class Config:
 
 		sections = ["General", "Backups", "IRC", "Death", "Proxy"]
 		defaults = {"General":{
-			"command": "java -jar minecraft_server.1.7.10.jar",
+			"command": "java -jar minecraft_server.1.8.jar",
 			"auto-restart": True,
 			"debug": False,
 			"pre-1.7-mode": False,
@@ -119,7 +120,8 @@ class Config:
 			"proxy-port": 25565,
 			"proxy-bind": "0.0.0.0",
 			"motd": "Minecraft Server",
-			"online-mode": True
+			"online-mode": True,
+			"max-players": 1024
 		}}
 		# Removed from the list above until fully implemented.
 		{"Web":{