+++ src/globals.py
@@ -1,2 +1,2 @@
-build=31
+build=37
 type='stable'